Clean up BenchmarkConfig (#168)
This PR does a lot of random clean ups.

The core changes are:
 - moved `BenchmarkConfig` to its own file
 - changed `write_ratio` to `operation`, with `Operation::Read` and `Operation::Write` (a pre-step of #165)
 - moved all utils into a `perma::utils` namespace
 - removed the `internal::` namespace, because it was annoying

There is basically no change in logic in this PR and the large file changes are just copy-paste from one to another location. So this can more or less just be skimmed to check for small mistakes. I'll check this myself again on Monday to have a small break before reviewing it again.  

This PR is based on #166, so that needs to be merged first. Overall, these are quite a few substantial refactorings, but hopefully we'll only need one or two PRs more.