Pre-generate addresses in benchmarks (#170)
This PR changes the core concept of how we execute benchmarks. Instead of pre-generating addresses in a small chunks at runtime, we pre-generate all data upfront to avoid the overhead of and gaps between these chunks. With the chunking approach, we overestimated the bandwidth which lead to results that were too high. 

The chunking approach still exists but only for time-based execution, i.e., when a benchmarks should end after X seconds. In that case, we need to periodically check where to end the benchmark or not. 

TODOs:
 - [x] better testing of time-based run
 - [x] check config validation in more detail to see if this is still correct for all cases
 - [x] clean up
 - [ ] per-thread information for custom ops (will do in #169)
 - [x] remove hacky `cat result_file` in `BenchmarkSuite.cpp`