Advanced custom ops (#175)
We can now specify whether a custom operation is in DRAM or PMem via `p`/`d` after a read `r` or write `w`. For example: `rp_64` (or as PMem is the default just `r_64`) is a 64 Byte read from PMem and `rd_128` is a 128 Byte read from DRAM.

Also, for write operations, we can now optionally specify the offset dependent on the previous read location via `_<offset>` at the end. For example: `w_64_cache_128` will perform a 64 Byte write 128 Bytes behind the previously read location. The idea is that we can "jump" around a bit. This is needed, e..g, when an index structure flushes data somewhere in the node and then updates the metadata at the beginning of the node. This can be modeled via the chain: `r_256,w_64_cache_128,w_64_cache_-128`. This will read 256 Bytes from PMem at address `100`. The first write will be done at address `100 + 128 = 228`, and the last write will be at address `228 - 128 = 100`, so back at the start. The offset is always relative to the previous operation. The offset is 0 by default and does not have to be specified.