Add number_partitions=0 (#178)
This PR allows users to set the number of partitions to 0. In this case, the number of partitions is set to the number of threads. This allows us to scale a partitioned workload with multiple threads. Before, the thread had to be fixed.

We can now do this for partitioned sequential writes, which was not possible before.
```yaml
seq_write:
  matrix:
    number_threads: [8, 16]
    number_partitions: [1, 2, 4, 8, 0]  # note 0 at end which will be 8 and 16

  args:
    access_size: 512
    memory_range: 10G
    persist_instruction: nocache
    operation: write
    exec_mode: sequential
```

`number_partitions = 0`  is now also the default setting, as it is more common for each thread to have its own region in sequential access and the impact of shared/disjoint regions on random access is negligible when the ranges are large enough.

Closes #173.